CentOS: Install openstack plugin package
Currently plugins and jobs for OpenStack environment are not included
by the base rally package but these are provided by the independent
plugin package. This change ensures the plugin package is installed.

Note that the plugin package is not available in Ubuntu/Debian now so
the package is still missing in these distributions.

Closes-Bug: #1969113
Change-Id: I7b3b285f1cb298bbabae012862ae5dc1223bf602